requirements.txt
psycopg[binary]==3.2.12
sentence-transformers==5.1.2
PyPDF2==3.0.1
faiss-cpu==1.9.0
openai==1.55.3
python-dotenv==1.1.1 
fastapi==0.115.5
uvicorn==0.32.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.3.0
psycopg-pool==3.2.7
